Move gardenlet identity creation to `controller` package
